Turned indicators into switching buttons
Indicators now switch between check and red slashed circle when clicked
